# Optimize createPeerIdFromKeypair