--- a/package.json
+++ b/package.json
@@ -34,6 +34,7 @@
   },
   "homepage": "https://github.com/ChainSafe/discv5#readme",
   "devDependencies": {
+    "@dapplion/benchmark": "^0.1.6",
     "@types/bn.js": "^4.11.5",
     "@types/chai": "^4.2.0",
     "@types/debug": "^4.1.5",