Seen in profiles
![createPeerIdFromKeypair](https://user-images.githubusercontent.com/1348242/126829615-1288683a-a1a9-4559-9c26-d4da1a2c1696.png)


Added benchmarks to show improvement:

Before:
```
  createPeerIdFromKeypair               
    ✓ createPeerIdFromKeypair - private key                               5892.995 ops/s    169.6930 us/op        -       4000 runs  0.683 s
    ✓ createPeerIdFromKeypair - no private key                            11393.80 ops/s    87.76700 us/op        -       4000 runs  0.354 s
```
After:
```
  createPeerIdFromKeypair                                                                      
    ✓ createPeerIdFromKeypair - private key                               10070.49 ops/s    99.30000 us/op        -       4000 runs  0.401 s
    ✓ createPeerIdFromKeypair - no private key                            25009.38 ops/s    39.98500 us/op        -       4000 runs  0.163 s
```